Auto update of GUI
Add the “update” method to the App class so that a user callback will
be executed after a specified time.
Example shows how user would specify a function to read a sensor at a
specified time interval and displayed on GUI. To make the example
portable, a random number is generated rather than reading a sensor.